CE-57 added a draft spike for replication

+++ src/main/java/net/openhft/chronicle/engine/tree/VanillaAssetTree.java
@@ -31,7 +31,7 @@ public class VanillaAssetTree implements AssetTree {
 
     }
 
-    public VanillaAssetTree(int hostId) {
+    public VanillaAssetTree(byte hostId) {
         root.addLeafRule(HostIdentifier.class, "host id holder", (rc, context) -> new HostIdentifier(hostId));
     }
